requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.3=py311h05b510d_1
aiosignal=1.3.1=pyhd8ed1ab_0
aiostream=0.5.2=pyhd8ed1ab_0
annotated-types=0.6.0=pyhd8ed1ab_0
anyio=4.3.0=pyhd8ed1ab_0
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
aws-c-auth=0.7.16=h51b92d1_6
aws-c-cal=0.6.10=hf888d4c_1
aws-c-common=0.9.13=h93a5062_0
aws-c-compression=0.2.18=hf888d4c_1
aws-c-event-stream=0.4.2=h1b95c93_4
aws-c-http=0.8.1=heb97226_5
aws-c-io=0.14.5=hb51f26e_1
aws-c-mqtt=0.10.2=hd37591b_4
aws-c-s3=0.5.2=h09f5dbc_0
aws-c-sdkutils=0.1.15=hf888d4c_1
aws-checksums=0.1.18=hf888d4c_1
aws-crt-cpp=0.26.2=h06549c9_7
aws-sdk-cpp=1.11.267=h73c0887_1
beautifulsoup4=4.12.3=pyha770c72_0
brotli-python=1.1.0=py311ha891d26_1
bzip2=1.0.8=h93a5062_5
c-ares=1.28.1=h93a5062_0
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
dataclasses-json=0.6.4=pyhd8ed1ab_0
datasets=2.19.0=pyhd8ed1ab_0
debugpy=1.8.1=py311h92babd0_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pyh1a96a4e_0
dill=0.3.8=pyhd8ed1ab_0
dirtyjson=1.0.8=pyhd8ed1ab_0
distro=1.9.0=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.4=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
frozenlist=1.4.1=py311h05b510d_0
fsspec=2024.3.1=pyhca7485f_0
gflags=2.2.2=hc88da5d_1004
glog=0.7.0=hc6770e3_0
gmp=6.3.0=hebf3989_1
gmpy2=2.1.2=py311h2ba9262_1
greenlet=3.0.3=py311h92babd0_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.5=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
huggingface_hub=0.22.2=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=hc8870d7_0
idna=3.7=pyhd8ed1ab_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.22.2=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.4.0=pyhd8ed1ab_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h267d04e_0
krb5=1.21.2=h92f50d5_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20230802.1=cxx17_h13dd4ca_0
libarrow=15.0.0=h4f70cd7_6_cpu
libarrow-acero=15.0.0=hebf3989_6_cpu
libarrow-dataset=15.0.0=hebf3989_6_cpu
libarrow-flight=15.0.0=hef2a8c6_6_cpu
libarrow-flight-sql=15.0.0=h611cb26_6_cpu
libarrow-gandiva=15.0.0=h5211482_6_cpu
libarrow-substrait=15.0.0=h2e851bb_6_cpu
libblas=3.9.0=22_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=22_osxarm64_openblas
libcrc32c=1.1.2=hbdafb3b_0
libcurl=8.7.1=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.20=h93a5062_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h2757513_1
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libgoogle-cloud=2.17.0=hfb9b341_2
libgoogle-cloud-storage=2.17.0=h8a76758_2
libgrpc=1.60.1=hfc68871_0
libiconv=1.17=h0d3ecfb_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=22_osxarm64_openblas
libllvm16=16.0.6=haab561b_3
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.27=openmp_h6c19121_0
libparquet=15.0.0=h278d484_6_cpu
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.1=h810fc01_2
libre2-11=2023.09.01=h741fcf5_1
libsentencepiece=0.1.99=h7c97dd4_6
libsodium=1.0.18=h27ca646_1
libsqlite=3.45.3=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.19.0=h026a170_1
libtiff=4.6.0=h07db509_3
libtorch=2.1.2=cpu_generic_hda0de89_3
libutf8proc=2.8.0=h1a8c8d9_0
libuv=1.48.0=h93a5062_0
libwebp-base=1.4.0=h93a5062_0
libxcb=1.15=hf346824_0
libxml2=2.12.6=h0d0cfa8_2
libzlib=1.2.13=h53f4e23_5
llvm-openmp=18.1.3=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
markupsafe=2.1.5=py311h05b510d_0
marshmallow=3.21.1=pyhd8ed1ab_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h41d338b_1
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.5=py311he2be06e_0
multiprocess=0.70.16=py311h05b510d_0
mypy_extensions=1.0.0=pyha770c72_0
ncurses=6.4.20240210=h078ce10_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pyhd8ed1ab_1
nltk=3.8.1=pyhd8ed1ab_0
nomkl=1.0=h5ca1d4c_0
numpy=1.26.4=py311h7125741_0
openai=1.23.1=pyhd8ed1ab_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.2.1=h0d3ecfb_1
orc=1.9.2=hb41d57e_1
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.2=py311hfbe21a1_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py311h0b5d0a1_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.8=py311h05b510d_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=15.0.0=py311hce53c6f_6_cpu
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pydantic=2.7.0=pyhd8ed1ab_0
pydantic-core=2.18.1=py311h94f323b_0
pygments=2.17.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.8=hdf0ec26_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python-xxhash=3.4.1=py311heffc1b2_0
python_abi=3.11=4_cp311
pytorch=2.1.2=cpu_generic_py311h38827da_3
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py311heffc1b2_1
pyzmq=26.0.0=py311he9c0408_0
re2=2023.09.01=h4cba328_1
readline=8.2=h92ec313_1
regex=2024.4.16=py311h05b510d_0
requests=2.31.0=pyhd8ed1ab_0
safetensors=0.4.3=py311h94f323b_0
scikit-learn=1.2.2=py311hf0b18b8_2
scipy=1.13.0=py311h4f9446f_0
sentence-transformers=2.7.0=pyhd8ed1ab_0
sentencepiece=0.1.99=ha1ab1f8_6
sentencepiece-python=0.1.99=py311ha79fcf5_6
sentencepiece-spm=0.1.99=h7c97dd4_6
setuptools=69.5.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
snappy=1.1.10=hd04f947_1
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
sqlalchemy=2.0.29=py311h05b510d_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypyh9d50eac_103
tenacity=8.2.3=pyhd8ed1ab_0
threadpoolctl=3.4.0=pyhc1e730c_0
tiktoken=0.6.0=py311hcd5b669_0
tk=8.6.13=h5083fa2_1
tokenizers=0.15.2=py311hcd5b669_0
torchvision=0.16.1=cpu_py311haf5c566_3
tornado=6.4=py311h05b510d_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.39.3=pyhd8ed1ab_0
typing-extensions=4.11.0=hd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
typing_inspect=0.9.0=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.43.0=pyhd8ed1ab_1
wrapt=1.16.0=py311h05b510d_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xxhash=0.8.2=hb547adb_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.4=py311h05b510d_0
zeromq=4.3.5=hebf3989_1
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.5=h4f39d0f_0
